port OPS to imglib2. mostly changed Cursor use to RandomAccess use and from Image use to Img use. note that I've hatched a class called RegionCursor to simplify the transition from imglib1 to imglib2. This class may not be efficient enough and should be replaced by an appropriate class (some kind of iterable rectangle region) from imglib2.roi when it is written.

+++ imglib2/ops/src/test/java/net/imglib2/ops/example/rev3/condition/And.java
@@ -13,7 +13,7 @@ public final class And implements Condition
 	}
 	
 	@Override
-	public boolean isSatisfied(IntegerIndexedScalarFunction function, int[] position)
+	public boolean isSatisfied(IntegerIndexedScalarFunction function, long[] position)
 	{
 		return left.isSatisfied(function, position) && right.isSatisfied(function, position);
 	}
