+++ imglib2/ops/src/test/java/net/imglib2/ops/example/rev3/condition/AxisGreaterThan.java
@@ -14,7 +14,7 @@ public final class AxisGreaterThan implements Condition
 	}
 	
 	@Override
-	public boolean isSatisfied(IntegerIndexedScalarFunction function, int[] position)
+	public boolean isSatisfied(IntegerIndexedScalarFunction function, long[] position)
 	{
 		return position[axis] > value;
 	}
