+++ imglib2/ops/src/test/java/net/imglib2/ops/example/rev3/condition/Not.java
@@ -12,7 +12,7 @@ public final class Not implements Condition
 	}
 	
 	@Override
-	public boolean isSatisfied(IntegerIndexedScalarFunction function, int[] position)
+	public boolean isSatisfied(IntegerIndexedScalarFunction function, long[] position)
 	{
 		return ! cond.isSatisfied(function, position);
 	}
