+++ imglib2/ops/src/test/java/net/imglib2/ops/example/rev3/constraints/Constraints.java
@@ -22,7 +22,7 @@ public class Constraints
 		constraints.add(new ConstraintEntry(function,condition));
 	}
 	
-	public boolean areSatisfied(int[] position)
+	public boolean areSatisfied(long[] position)
 	{
 		if (constraints == null)
 			return true;
