+++ imglib2/ops/src/test/java/net/imglib2/ops/example/rev3/function/ConstantFunction.java
@@ -11,7 +11,7 @@ public final class ConstantFunction implements IntegerIndexedScalarFunction
 	}
 	
 	@Override
-	public double evaluate(int[] position)
+	public double evaluate(long[] position)
 	{
 		return value;
 	}
